+++ autcli/utils.py
@@ -307,6 +307,16 @@ def parse_token_value_representation(value_str: str, decimals: int) -> int:
     return int(Decimal(value_str) * Decimal(pow(10, decimals)))
 
 
+def format_quantity(units: int, decimals: int) -> str:
+    """
+    Given some quantity of atomic "token units" of a currency
+    (e.g. Attoton for Auton), and the decimals used to represent the
+    value of such a unit, return a string representation of the number
+    of tokens.
+    """
+    return str(Decimal(units) * Decimal(pow(10, decimals)))
+
+
 def address_keyfile_dict(keystore_dir: str) -> Dict[ChecksumAddress, str]:
     """
     For directory 'keystore' that contains one or more keyfiles,
@@ -443,6 +453,20 @@ def newton_or_token_to_address(
         return Web3.toChecksumAddress(token)
 
     return None
+
+
+def newton_or_token_to_address_require(
+    ntn: bool, token: Optional[str]
+) -> ChecksumAddress:
+    """
+    Similar to newton_or_token_address, but thrown an error if neither
+    is given.
+    """
+    token = newton_or_token_to_address(ntn, token)
+    if token is None:
+        raise ClickException("Token address (or --ntn) must be specified.")
+
+    return token
 
 
 def prompt_for_new_password(show_password: bool) -> str: